saturday 09:28

+++ Triangle.java
@@ -6,9 +6,9 @@
 public class Triangle {
 	public static void main(String[] args) {
 		// Put your code here	
-		double a = Integer.parseInt(args[0]);
-        double b = Integer.parseInt(args[1]);
-        double c = Integer.parseInt(args[2]);
+		  int a = Integer.parseInt(args[0]);
+        int b = Integer.parseInt(args[1]);
+        int c = Integer.parseInt(args[2]);
         boolean istriangle ;
 
         if ((a + b) > c & (b + c) > a & (a + c)> b) {
